[chore] 친구 관련 기능은 friendController 로 분리 (#LYC-76)
